refactoring: move DefaultNearMethods to io.emeraldpay.dshackle.upstream.calls
